+++ src/pages/auth/auth.css
@@ -128,6 +128,10 @@ input[type='checkbox']:checked + label {
 /* Components */
 
 /* Login form */
+.login-session {
+  display: block;
+}
+
 #emailInput[type='email'] {
   width: 100%;
   margin-top: 57px;
@@ -189,6 +193,10 @@ input[type='checkbox']:checked + label {
 }
 
 /* Password form */
+.password-session {
+  display: none;
+}
+
 .edit-container {
   display: flex;
 }
@@ -218,7 +226,7 @@ input[type='checkbox']:checked + label {
 }
 
 #regContainer {
-  min-width: 360px;
+  min-width: 300px;
   height: 40px;
   background-color: rgb(231, 229, 229);
   margin-top: 20px;
