+++ vite.config.js
@@ -8,10 +8,9 @@ export default defineConfig({
         footer: 'src/components/footer/footer.html',
         index: 'index.html', // 기본 index.html
         sign: 'src/pages/auth/signup.html',
-        login: 'src/pages/auth/login.html', // 추가 HTML 파일
-        pw: 'src/pages/auth/pw.html',
+        login: 'src/pages/auth/login.html',
         check: 'src/pages/auth/check.html',
-        list: 'src/pages/product/list/index.html', // 추가 HTML 파일
+        list: 'src/pages/product/list/index.html',
       },
     },
   },
